pyproject.toml
[project]
name = "us-visa"
version = "0.1.0"
description = ""
authors = [
    {name = "Stavan",email = "stavan.sanyal@gmail.com"}
]
readme = "README.md"
requires-python = ">3.10, <3.12"
dependencies = [
    "ipykernel (>=6.29.5,<7.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "plotly (>=6.0.1,<7.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "imblearn (>=0.0,<0.1)",
    "xgboost (>=3.0.0,<4.0.0)",
    "catboost (>=1.2.8,<2.0.0)",
    "pymongo (>=4.12.0,<5.0.0)",
    "from-root (>=1.3.0,<2.0.0)",
    "evidently (==0.2.8)",
    "dill (>=0.3.9,<0.4.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "neuro-mf (>=0.0.5,<0.0.6)",
    "boto3 (>=1.37.33,<2.0.0)",
    "mypy-boto3-s3 (>=1.37.24,<2.0.0)",
    "botocore (>=1.37.33,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
